+++ examples/with_context/s3_bucket/s3_bucket/__main__.py
@@ -1,0 +1,4 @@
+"""Allow running as: python -m s3_bucket."""
+from .main import main
+
+main()